[ring] : re-organisation of files + test revamp
