+++ ring/ring_automorphism.go
