+++ ring/ring_ntt.go
