+++ ring/ring_poly.go
