+++ ring/ring_sampler_ternary.go
