+++ ring/ring_sampler_uniform.go
